나홀로 백엔드 4일차...

+++ src/main/java/com/likelionskhu/hagseubjang/service/WishService.java
@@ -24,7 +24,15 @@ public class WishService {
     private final HttpSession httpSession;
 
     @Transactional
-    public void save(int lectureId) {
+    public void delete(int id) {
+        Wish wish = wishRepository.findById(id)
+                .orElseThrow(() -> new IllegalArgumentException("해당 찜이 없습니다. id=" + id));
+
+        wishRepository.delete(wish);
+    }
+
+    @Transactional
+    public void saveOrDelete(int lectureId) {
         SessionUser user = (SessionUser) httpSession.getAttribute("user");
         User user1 = userRepository.findByEmail(user.getEmail())
                 .orElseThrow(() -> new IllegalArgumentException("해당 유저가 없습니다. email=" + user.getEmail()));
@@ -32,18 +40,16 @@ public class WishService {
         Lecture lecture = lectureRepository.findById(lectureId)
                 .orElseThrow(() -> new IllegalArgumentException("해당 강좌가 없습니다. id=" + lectureId));
 
-        Wish wish = new Wish();
-        wish.setUser(user1);
-        wish.setLecture(lecture);
-        wishRepository.save(wish);
-    }
+        Wish wish = wishRepository.findByLectureAndUser(lecture, user1);
 
-    @Transactional
-    public void delete(int id) {
-        Wish wish = wishRepository.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("해당 찜이 없습니다. id=" + id));
-
-        wishRepository.delete(wish);
+        if (wish == null) {
+            wish = new Wish();
+            wish.setLecture(lecture);
+            wish.setUser(user1);
+            wishRepository.save(wish);
+        } else {
+            wishRepository.delete(wish);
+        }
     }
 
 }
